refactor: 팔로잉시 팔로잉 수가 늘어나지 않는 문제 해결
- 엔티티에서 잘못된 변수로 인해 팔로잉 수가 늘어나지 않던 것 변경
